Function-ised finding the tag
Finding the tag (and loading the snippet contents) is now in its
own function.  This should make it easier to swap in a different
storage mechanism for the snippets.

There's still a couple of memory issues to deal with, and a few
places where buffer sizes aren't checked.

Note to maintainer: The conversion from ANSI to UTF-8 needs looking
at - the buffer sizes are wrong I think - the result could be bigger
than the original. It's kinda worse now, in that the buffer is now
the right size, so if it is bigger, it definitely won't fit, but it
wasn't checked before either.