fixed from PVS-Studio:
V575 The number of processed elements should be passed to the 'GetWindowTextW' function as the third argument instead of buffer's size in bytes. dockingdlginterface.h 43
V646 Consider inspecting the application's logic. It's possible that 'else' keyword is missing. plugindefinition.cpp 1695
V668 There is no sense in testing the 'pszOutBuffer' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. plugindefinition.cpp 5615
V773 Visibility scope of the 'tempScope' pointer was exited without releasing the memory. A memory leak is possible. plugindefinition.cpp 593